// src/types/interface.ts
export interface Contact {
  id: string;
  createdate: string;
  properties: {
    createdate: string;
    hs_object_id: string;
    firstname: string;
    lastname: string;
    email: string;
    phone: string;
    hs_v2_date_entered_customer: string;
    hs_v2_date_entered_lead: string;
    lastmodifieddate: string;
  };
}